inserindo vagas no banco e atualizando informaçoes para inserçao em b…

+++ src/main/java/br/com/temvaga/repository/UsuarioRepository.java
@@ -4,8 +4,9 @@ import br.com.temvaga.model.Usuario;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.ArrayList;
+import java.util.Optional;
 
-public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
+public interface UsuarioRepository extends CrudRepository<Usuario, Integer> {
 
     ArrayList<Usuario> findAll();
